backlog - 50%

+++ app/Http/Controllers/ScrumController.php
@@ -11,9 +11,11 @@ use App\Models\Activity;
 use App\Models\MeetingParticipants;
 use App\Models\ProjectMembers;
 use Carbon\Carbon;
-use App\Models\Epic;
+use App\Models\Epics;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
+use App\Models\Sprints;
+
 
 class ScrumController extends Controller
 {
@@ -97,22 +99,35 @@ class ScrumController extends Controller
         ]);
     }
     public function getBacklogDatas(Request $request) {
-        $projectDetails = Project::where('id', $request->query('id'))->first();
-
-        $epics = Epic::where('project_id', $projectDetails->id)->get();
+        if ($request->query('id')) {
+            $projectDetails = Project::where('id', $request->query('id'))->first();
+        } else {
+            $projectDetails = Project::where('id', $request->projectId)->first();
+        }
+        $sprint = [];
+        $epics = Epics::where('project_id', $projectDetails->id)->get();
         $backlogs = Backlogs::where('project_id', $projectDetails->id)->get();
+        foreach ($epics as $epic) {
+            if ($epic->status == "On Sprint") {
+                $sprint = Sprints::where('epic_id', $epic->id)
+                ->where('status', 'Active')
+                ->first();
+            }
+        }
         return Inertia::render('Scrum/ScrumBacklog', [
             'projectDetails' => $projectDetails,
             'backlogs' => $backlogs,
             'epics' => $epics,
+            'sprint' => $sprint,
         ]);
     }
 
     public function updateEpicOrder(Request $request) {
+        Log::info('Received epics data:', ['epics' => $request->epics]);
         $epics = $request->epics;
         
         foreach($epics as $index => $epic) {
-            Epic::where('id', $epic['epic_id'])
+            Epics::where('id', $epic['epic_id'])
                 ->update(['order' => $index + 1]);
         }
         return response()->json(['success' => true]);
@@ -120,16 +135,18 @@ class ScrumController extends Controller
 
     public function createEpic(Request $request) {
         $projId = $request->projectId;
-        $epicNum = Epic::where('project_id', $projId)->count() + 1;
-        Epic::create([
+        $epicNum = Epics::where('project_id', $projId)->count();
+        $epic = Epics::create([
             'project_id' => $projId,
             'name' => $request->name,
             'description' => '',
-            'progress_precent' => 0,
+            'progress_percent' => 0,
+            'status' => 'Pending',
             'key' => $request->key,
             'order' => $epicNum + 1,
         ]);
-        return response()->json(['success' => true]);
+        $this->registerUpdate($projId, Auth::user()->id, ' created an epic named ', $request->name);
+        return response()->json(['success' => true, 'id' => $epic->id]);
     }
 
     public function createBacklog(Request $request) {
@@ -137,32 +154,35 @@ class ScrumController extends Controller
         $backlogNum = Backlogs::where('project_id', $projId)
         ->where('epic_id', $request->epicId)
         ->count() + 1;
-        $epic = Epic::where('id', $request->epicId)->first();
-        $backlogCreated = Backlogs::create([
+        $epic = Epics::where('id', $request->epicId)->first();
+        $creator = ProjectMembers::where('user_id', Auth::user()->id)->value('id');
+        $backlog = Backlogs::create([
             'title' => $request->title,
             'project_id' => $projId,
             'type' => $request->type,
             'description' => '',
             'priority' => $request->priority,
             'epic_id' => $request->epicId,
-            'creator_id' => Auth::user()->id,
+            'creator_id' => $creator,
             'status' => 'To Do',
             'order' => $request->order,
         ]);
         $this->registerUpdate($projId, Auth::user()->id, "created " . $request->title . " in ", $epic->name);
-        return response()->json(['success' => true, 'id' => $backlogCreated->id]);
+        return response()->json(['success' => true, 'id' => $backlog->id]);
     }
 
     public function deleteBacklog(Request $request) {
         $backlog = Backlogs::where('id', $request->id)->first();
-        $backlog->delete();
-        $epic = Epic::where('id', $request->epicId)->first();
-        $desc = "deleted " . $request->title . " in ";
-        $updateResult = $this->registerUpdate($request->projectId, Auth::user()->id, $desc, $epic->name);
-        if ($updateResult) {
-            return response()->json(['success' => true]);;
+        if ($backlog) {
+            $epic = Epics::where('id', $request->epicId)->first();
+            $desc = "deleted " . $request->title . " in ";
+            $backlog->delete();
+            $updateResult = $this->registerUpdate($request->projectId, Auth::user()->id, $desc, $epic->name);
+            return response()->json(['success' => true]);
         }
+        return response()->json(['success' => false], 404);
     }
+    
 
     private function registerUpdate($projectId ,$userId, $description,$subject) {
         try {
@@ -183,6 +203,52 @@ class ScrumController extends Controller
             ]);
             throw $e;
         }
+    }
+
+    public function updateBacklogStatus(Request $request) {
+        $backlog = Backlogs::where('id', $request->id)->first();
+        $backlog->status = $request->status;
+        $backlog->save();
+        
+        $epic = Epics::where('id', $backlog->epic_id)->first();
+        $this->registerUpdate($request->projectId, Auth::user()->id, "updated status of " . $backlog->title . " to " . $request->status . " in ", $epic->name);
+        
+        return response()->json(['success' => true]);
+    }
+
+    public function updateEpicStatus(Request $request) {
+        $epic = Epics::where('id', $request->epicId)->first();
+        $epic->status = $request->status;
+        $epic->save();
+        
+        $this->registerUpdate(
+            $request->projectId, 
+            Auth::user()->id, 
+            "updated epic status to " . $request->status . " for ", 
+            $epic->name
+        );
+        
+        return response()->json(['success' => true]);
+    }
+    
+    public function startSprint(Request $request) {
+        $request->validate([
+            'projectId' => 'required|integer',
+            'name' => 'required|string|max:255',
+            'start_date' => 'required|date',
+            'end_date' => 'required|date|after:start_date',
+        ]);
+        $epic = Epics::where('id', $request->epic_id)->first();
+        $sprint = Sprints::create([
+            'name' => $request->name,
+            'epic_id' => $request->epic_id,
+            'status' => 'Active',
+            'start_date' => $request->start_date,
+            'end_date' => $request->end_date,
+            'description' => $request->description,
+        ]);
+        $this->registerUpdate($request->projectId, Auth::user()->id, "started a sprint named " . $request->name, $epic->name);
+        return redirect()->back()->with('success', 'Sprint started successfully.');
     }
 }
